for_1.2.0 QA 수정 내용 반영 2

+++ app/src/main/java/com/ham/onettsix/adapter/NoticeAdapter.kt
@@ -23,7 +23,7 @@ class NoticeAdapter : RecyclerView.Adapter<NoticeAdapter.NoticeViewHolder>() {
             binding.noticeItemTitleTv.text = item.title
             binding.noticeItemContentTv.text = item.content
             binding.noticeItemTimeTv.text =
-                TimeUtils.timeDiff(itemView.context, item.createdAt.toLong())
+                TimeUtils.timeDiff(itemView.context, item.createdAt)
             Glide.with(itemView.context).load(R.drawable.ic_notifications)
                 .into(binding.noticeItemIconImg)
         }
